Add an 'evaluate' operation to TemplateExpression elements
Every TemplateExpression element now has an 'evaluate' operation that
accepts a list of Assignment elements. This operation returns the value
that the expression assumes with the given assignment of parameters.
The implementation of such operation is given as OCL queries and the Java
code is then automaticcaly generated by EMF.